[ReportBundle] CS fixes
