+++ src/Sylius/Bundle/ReportBundle/Form/EventListener/BuildReportDataFetcherFormSubscriber.php
@@ -1,18 +1,17 @@
 <?php
-
 /*
- * This file is part of the Sylius package.
- *
- * (c) Paweł Jędrzejewski
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
+* This file is part of the Sylius package.
+*
+* (c) Paweł Jędrzejewski
+*
+* For the full copyright and license information, please view the LICENSE
+* file that was distributed with this source code.
+*/
 
 namespace Sylius\Bundle\ReportBundle\Form\EventListener;
 
-use Sylius\Component\Report\Model\ReportInterface;
 use Sylius\Component\Registry\ServiceRegistryInterface;
+use Sylius\Component\Report\Model\ReportInterface;
 use Sylius\Component\Resource\Exception\UnexpectedTypeException;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Form\FormEvent;
@@ -22,33 +21,36 @@ use Symfony\Component\Form\FormInterface;
 
 /**
  * This listener adds configuration form to the report object
- * if selected renderer requires one.
+ * if selected data fetcher requires one.
  *
- * @author Mateusz Zalewski <mateusz.zalewski@lakion.com>
+ * @author Łukasz Chruściel <lukasz.chrusciel@lakion.com>
  */
-class BuildReportRendererFormListener implements EventSubscriberInterface
+class BuildReportDataFetcherFormSubscriber implements EventSubscriberInterface
 {
     /**
      * @var ServiceRegistryInterface
      */
-    private $rendererRegistry;
+    private $dataFecherRegistry;
 
     /**
      * @var FormFactoryInterface
      */
     private $factory;
 
-    public function __construct(ServiceRegistryInterface $rendererRegistry, FormFactoryInterface $factory)
+    public function __construct(ServiceRegistryInterface $dataFecherRegistry, FormFactoryInterface $factory)
     {
-        $this->rendererRegistry = $rendererRegistry;
+        $this->dataFecherRegistry = $dataFecherRegistry;
         $this->factory = $factory;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public static function getSubscribedEvents()
     {
         return array(
             FormEvents::PRE_SET_DATA => 'preSetData',
-            FormEvents::PRE_SUBMIT   => 'preBind',
+            FormEvents::PRE_SUBMIT => 'preBind',
         );
     }
 
@@ -64,37 +66,37 @@ class BuildReportRendererFormListener implements EventSubscriberInterface
             throw new UnexpectedTypeException($report, 'Sylius\Component\Report\Model\ReportInterface');
         }
 
-        $this->addConfigurationFields($event->getForm(), $report->getRenderer(), $report->getRendererConfiguration());
+        $this->addConfigurationFields($event->getForm(), $report->getDataFetcher(), $report->getDataFetcherConfiguration());
     }
 
     public function preBind(FormEvent $event)
     {
         $data = $event->getData();
 
-        if (empty($data) || !array_key_exists('renderer', $data)) {
+        if (empty($data) || !array_key_exists('dataFetcher', $data)) {
             return;
         }
 
-        $this->addConfigurationFields($event->getForm(), $data['renderer']);
+        $this->addConfigurationFields($event->getForm(), $data['dataFetcher']);
     }
 
     /**
-     * Add configuration fields to the form
+     * Add configuration fields to the form.
      *
      * @param FormInterface $form
-     * @param string        $rendererType
-     * @param array         $data
+     * @param string        $dataFetcherType
+     * @param array         $config
      */
-    public function addConfigurationFields(FormInterface $form, $rendererType, array $data = array())
+    protected function addConfigurationFields(FormInterface $form, $dataFetcherType, array $config = array())
     {
-        $renderer = $this->rendererRegistry->get($rendererType);
-        $formType = sprintf('sylius_renderer_%s', $renderer->getType());
+        $dataFetcher = $this->dataFecherRegistry->get($dataFetcherType);
+        $formType = sprintf('sylius_data_fetcher_%s', $dataFetcher->getType());
 
         try {
             $configurationField = $this->factory->createNamed(
-                'rendererConfiguration',
+                'dataFetcherConfiguration',
                 $formType,
-                $data,
+                $config,
                 array('auto_initialize' => false)
             );
         } catch (\InvalidArgumentException $e) {
