+++ src/Sylius/Bundle/ReportBundle/Form/EventListener/BuildReportRendererFormSubscriber.php
@@ -1,17 +1,18 @@
 <?php
+
 /*
-* This file is part of the Sylius package.
-*
-* (c) Paweł Jędrzejewski
-*
-* For the full copyright and license information, please view the LICENSE
-* file that was distributed with this source code.
-*/
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 namespace Sylius\Bundle\ReportBundle\Form\EventListener;
 
-use Sylius\Component\Registry\ServiceRegistryInterface;
 use Sylius\Component\Report\Model\ReportInterface;
+use Sylius\Component\Registry\ServiceRegistryInterface;
 use Sylius\Component\Resource\Exception\UnexpectedTypeException;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Form\FormEvent;
@@ -21,36 +22,33 @@ use Symfony\Component\Form\FormInterface;
 
 /**
  * This listener adds configuration form to the report object
- * if selected data fetcher requires one.
+ * if selected renderer requires one.
  *
- * @author Łukasz Chruściel <lukasz.chrusciel@lakion.com>
+ * @author Mateusz Zalewski <mateusz.zalewski@lakion.com>
  */
-class BuildReportDataFetcherFormListener implements EventSubscriberInterface
+class BuildReportRendererFormSubscriber implements EventSubscriberInterface
 {
     /**
      * @var ServiceRegistryInterface
      */
-    private $dataFecherRegistry;
+    private $rendererRegistry;
 
     /**
      * @var FormFactoryInterface
      */
     private $factory;
 
-    public function __construct(ServiceRegistryInterface $dataFecherRegistry, FormFactoryInterface $factory)
+    public function __construct(ServiceRegistryInterface $rendererRegistry, FormFactoryInterface $factory)
     {
-        $this->dataFecherRegistry = $dataFecherRegistry;
+        $this->rendererRegistry = $rendererRegistry;
         $this->factory = $factory;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public static function getSubscribedEvents()
     {
         return array(
             FormEvents::PRE_SET_DATA => 'preSetData',
-            FormEvents::PRE_SUBMIT => 'preBind',
+            FormEvents::PRE_SUBMIT   => 'preBind',
         );
     }
 
@@ -66,37 +64,37 @@ class BuildReportDataFetcherFormListener implements EventSubscriberInterface
             throw new UnexpectedTypeException($report, 'Sylius\Component\Report\Model\ReportInterface');
         }
 
-        $this->addConfigurationFields($event->getForm(), $report->getDataFetcher(), $report->getDataFetcherConfiguration());
+        $this->addConfigurationFields($event->getForm(), $report->getRenderer(), $report->getRendererConfiguration());
     }
 
     public function preBind(FormEvent $event)
     {
         $data = $event->getData();
 
-        if (empty($data) || !array_key_exists('dataFetcher', $data)) {
+        if (empty($data) || !array_key_exists('renderer', $data)) {
             return;
         }
 
-        $this->addConfigurationFields($event->getForm(), $data['dataFetcher']);
+        $this->addConfigurationFields($event->getForm(), $data['renderer']);
     }
 
     /**
      * Add configuration fields to the form.
      *
      * @param FormInterface $form
-     * @param string        $dataFetcherType
-     * @param array         $config
+     * @param string        $rendererType
+     * @param array         $data
      */
-    protected function addConfigurationFields(FormInterface $form, $dataFetcherType, array $config = array())
+    public function addConfigurationFields(FormInterface $form, $rendererType, array $data = array())
     {
-        $dataFetcher = $this->dataFecherRegistry->get($dataFetcherType);
-        $formType = sprintf('sylius_data_fetcher_%s', $dataFetcher->getType());
+        $renderer = $this->rendererRegistry->get($rendererType);
+        $formType = sprintf('sylius_renderer_%s', $renderer->getType());
 
         try {
             $configurationField = $this->factory->createNamed(
-                'dataFetcherConfiguration',
+                'rendererConfiguration',
                 $formType,
-                $config,
+                $data,
                 array('auto_initialize' => false)
             );
         } catch (\InvalidArgumentException $e) {
