+++ src/Sylius/Bundle/ReportBundle/Resources/config/serializer/Model.Report.yml
