+++ src/Sylius/Bundle/ReportBundle/Resources/views/Chart/default.html.twig
@@ -84,7 +84,7 @@
         }
         function generateRadialData(labels, values) {
             var colors = ["#1abb9c", "#F7464A", "#46BFBD", "#FDB45C", "#949FB1", "#4D5360", "#bdbdbd", "#AA66CC", "33B5E5"];
-            var highlights = ["#37a37e", "#FF5A5E", "#5AD3D1", "#FFC870", "#A8B3C5", "#616774", "#acacac", "#9933CC", "0099CC"]; 
+            var highlights = ["#37a37e", "#FF5A5E", "#5AD3D1", "#FFC870", "#A8B3C5", "#616774", "#acacac", "#9933CC", "0099CC"];
             var data = [];
             $.each(values, function(i){
                 var fragment = {
